+++ examples/data-sources/nxos_isis_instance/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_isis_instance" "example" {
+  name = "ISIS1"
+}