+++ examples/resources/nxos_isis/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_isis.example "sys/isis"