+++ examples/resources/nxos_isis_instance/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_isis_instance.example "sys/isis/inst-[ISIS1]"