+++ examples/resources/nxos_isis_instance/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_isis_instance" "example" {
+  name        = "ISIS1"
+  admin_state = "enabled"
+}