+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - Add nxos_feature_vn_segment resource and data source
 - Add nxos_feature_nv_overlay resource and data source
 - Add nxos_system resource and data source
+- Add nxos_isis resource and data source
+- Add nxos_isis_instance resource and data source
 
 ## 0.3.0